Add files via upload
Fuck it, let's just put Python in the game. I think this will work, we'll probably just need to right click the .py file and click "open with", and then select this. Worst case scenario, we have to get these files to C:\Users\[User]\Appdata\Local\Python. Which will be impossible.